requirements.txt
playwright
python-dotenv
geopy
starlink-grpc-core
yagmail
cryptography
rich
inquirer
